Clean up comments, unused code
